# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Waelio Marketing API
  description: REST API for Waelio advertising platform
  version: 1.0.0
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://ads.waelio.com
    description: Production ad server

tags:
  - name: Auth
  - name: Ads
  - name: Campaigns
  - name: Admin

paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK

  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                firstName: { type: string }
                lastName: { type: string }
                role: { type: string, enum: [ADVERTISER, PUBLISHER] }
      responses:
        '201':
          description: Created

  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Tokens returned

  /api/auth/logout:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: Logout
      responses:
        '200':
          description: Logged out

  /api/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset
      responses:
        '200':
          description: Email sent if exists

  /api/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password with token
      responses:
        '200':
          description: Password updated

  /api/ads:
    get:
      tags: [Ads]
      summary: Fetch ad for placement
      parameters:
        - name: placementId
          in: query
          required: true
          schema: { type: string }
        - name: device
          in: query
          schema: { type: string }
        - name: browser
          in: query
          schema: { type: string }
        - name: country
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Ad payload
        '204':
          description: No fill

  /api/impression:
    post:
      tags: [Ads]
      summary: Record impression
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                impressionId: { type: string }
      responses:
        '200':
          description: Recorded

  /api/click:
    post:
      tags: [Ads]
      summary: Record click
      responses:
        '200':
          description: Recorded

  /api/campaigns:
    get:
      tags: [Campaigns]
      security: [{ bearerAuth: [] }]
      summary: List campaigns
      responses:
        '200':
          description: Campaign list
    post:
      tags: [Campaigns]
      security: [{ bearerAuth: [] }]
      summary: Create campaign
      responses:
        '201':
          description: Created

  /api/campaigns/{id}/pause:
    post:
      tags: [Campaigns]
      security: [{ bearerAuth: [] }]
      summary: Pause campaign
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Paused

  /api/campaigns/{id}/resume:
    post:
      tags: [Campaigns]
      security: [{ bearerAuth: [] }]
      summary: Resume campaign
      responses:
        '200':
          description: Resumed

  /api/admin/metrics:
    get:
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      summary: System metrics
      responses:
        '200':
          description: Metrics

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT